Split LaunchTestIT into LaunchTestRunToolIT and LaunchTestYamlIT
